refactor: extract shared handler
