+++ local-worker.js
@@ -1,96 +1,9 @@
-const express = require('express');
-const fs = require('fs/promises');
+import express from 'express';
+import fs from 'fs/promises';
+import { handleRequest } from './handler.js';
 
 const app = express();
 app.use(express.json());
-
-const corsHeaders = {
-  'Access-Control-Allow-Origin': '*',
-  'Access-Control-Allow-Methods': 'GET,POST,OPTIONS',
-  'Access-Control-Allow-Headers': 'Content-Type'
-};
-
-async function handleRequest(request, env) {
-  const { method } = request;
-  const url = new URL(request.url);
-
-  if (method === 'OPTIONS') {
-    return new Response(null, { headers: corsHeaders });
-  }
-
-  if (url.pathname === '/orders') {
-    switch (method) {
-        case 'GET': {
-          let data = await env.ORDERS.get('list');
-          if (data === null) {
-            data = '[]';
-            await env.ORDERS.put('list', data);
-          }
-          return new Response(data, {
-            headers: { ...corsHeaders, 'Content-Type': 'application/json' }
-          });
-        }
-      case 'POST': {
-        let body;
-        try {
-          body = await request.json();
-        } catch {
-          return new Response(JSON.stringify({ error: 'Invalid JSON' }), {
-            status: 400,
-            headers: { ...corsHeaders, 'Content-Type': 'application/json' }
-          });
-        }
-        let list = await env.ORDERS.get('list', 'json');
-        if (!Array.isArray(list)) list = [];
-        list.push(body);
-        await env.ORDERS.put('list', JSON.stringify(list, null, 2));
-        return new Response(JSON.stringify({ status: 'ok' }), {
-          headers: { ...corsHeaders, 'Content-Type': 'application/json' }
-        });
-      }
-      default:
-        return new Response('Method Not Allowed', {
-          status: 405,
-          headers: corsHeaders
-        });
-    }
-  } else if (url.pathname === '/page_content.json') {
-    switch (method) {
-        case 'GET': {
-          let data = await env.PAGE_CONTENT.get('data');
-          if (data === null) {
-            data = '{}';
-            await env.PAGE_CONTENT.put('data', data);
-          }
-          return new Response(data, {
-            headers: { ...corsHeaders, 'Content-Type': 'application/json' }
-          });
-        }
-      case 'POST': {
-        try {
-          await request.clone().json();
-        } catch {
-          return new Response(JSON.stringify({ error: 'Invalid JSON' }), {
-            status: 400,
-            headers: { ...corsHeaders, 'Content-Type': 'application/json' }
-          });
-        }
-        const text = await request.text();
-        await env.PAGE_CONTENT.put('data', text);
-        return new Response(JSON.stringify({ status: 'ok' }), {
-          headers: { ...corsHeaders, 'Content-Type': 'application/json' }
-        });
-      }
-      default:
-        return new Response('Method Not Allowed', {
-          status: 405,
-          headers: corsHeaders
-        });
-    }
-  }
-
-  return new Response('Not Found', { status: 404, headers: corsHeaders });
-}
 
 const env = {
   ORDERS: {
