+++ package.json
@@ -13,5 +13,6 @@
   },
   "dependencies": {
     "express": "^5.1.0"
-  }
+  },
+  "type": "module"
 }
